+++ CHANGELOG.rst
@@ -2,7 +2,7 @@
 CHANGELOG
 =========
 
-Unreleased
+1.3.0
 ==========
 
 * Added: Support for explicit proxy configuration.